@@ -1,5 +1,5 @@
-export const UserPoolId = 'us-west-2_ANuvKCOIC';
-export const UserPoolClientId = '3d2aetrrr41j7t7f29oe7ccclo';
-export const IdentityPoolId = 'us-west-2:b9753997-53af-4ec4-853f-d3ad65332d0a';
+export const UserPoolId = 'us-west-2_npfbMYPty';
+export const UserPoolClientId = '1su93kh8al3mshji7ai6kqs2ih';
+export const IdentityPoolId = 'us-west-2:c47c5601-6684-4b90-bf27-1a0e4afff79a';
 export const CognitoEndpoint = 'cognito-idp.us-west-2.amazonaws.com';
 export const LoginEndpoint = `${CognitoEndpoint}/${UserPoolId}`;
